--- src/main/java/faang/school/notificationservice/config/redis/RedisListenerConfig.java
@@ -1,42 +1,0 @@
-package faang.school.notificationservice.config.redis;
-
-import faang.school.notificationservice.listener.RedisContainerMessageListener;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-
-import java.util.List;
-
-@Configuration
-@RequiredArgsConstructor
-public class RedisListenerConfig {
-
-    private final List<RedisContainerMessageListener> listeners;
-
-    @Value("${spring.data.redis.port}")
-    private int port;
-
-    @Value("${spring.data.redis.host}")
-    private String host;
-
-    @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(host, port);
-        return new JedisConnectionFactory(config);
-    }
-
-    @Bean
-    public RedisMessageListenerContainer container() {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(redisConnectionFactory());
-        listeners.forEach(listener ->
-                container.addMessageListener(listener.getAdapter(), listener.getTopic()));
-
-        return container;
-    }
-}
