+++ src/main/java/faang/school/notificationservice/config/vonage/VonageConfig.java
@@ -1,0 +1,24 @@
+package faang.school.notificationservice.config.vonage;
+
+import com.vonage.client.VonageClient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class VonageConfig {
+
+    @Value("${vonage.api.key}")
+    private String apiKey;
+
+    @Value("${vonage.api.secret}")
+    private String apiSecret;
+
+    @Bean
+    public VonageClient vonageClient() {
+        return VonageClient.builder()
+                .apiKey(apiKey)
+                .apiSecret(apiSecret)
+                .build();
+    }
+}
