--- src/main/java/faang/school/notificationservice/dto/EventTimeToStart.java
@@ -1,9 +1,0 @@
-package faang.school.notificationservice.dto;
-
-public enum EventTimeToStart {
-    START_NOW,
-    START_THROUGH_10_MIN,
-    START_THROUGH_1_HOUR,
-    START_THROUGH_5_HOURS,
-    START_THROUGH_1_DAY
-}
