--- src/main/java/faang/school/notificationservice/dto/event/AchievementEvent.java
@@ -1,16 +1,0 @@
-package faang.school.notificationservice.dto.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AchievementEvent {
-    private Long userId;
-    private String title;
-    private String description;
-}
