--- src/main/java/faang/school/notificationservice/event/MentorshipAcceptedEvent.java
@@ -1,19 +1,0 @@
-package faang.school.notificationservice.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class MentorshipAcceptedEvent {
-
-    private long id;
-    private long requesterId;
-    private String receiverUsername;
-}
