--- src/main/java/faang/school/notificationservice/event/UserFollowerEvent.java
@@ -1,9 +1,0 @@
-package faang.school.notificationservice.event;
-
-import lombok.Data;
-
-@Data
-public class UserFollowerEvent {
-    private long followerId;
-    private long followeeId;
-}
