+++ src/main/java/faang/school/notificationservice/event/event_start/EventTimeToStart.java
@@ -1,0 +1,9 @@
+package faang.school.notificationservice.event.event_start;
+
+public enum EventTimeToStart {
+    START_NOW,
+    START_THROUGH_10_MIN,
+    START_THROUGH_1_HOUR,
+    START_THROUGH_5_HOURS,
+    START_THROUGH_1_DAY
+}
