+++ src/main/java/faang/school/notificationservice/event/follower/UserFollowerEvent.java
@@ -1,0 +1,9 @@
+package faang.school.notificationservice.event.follower;
+
+import lombok.Data;
+
+@Data
+public class UserFollowerEvent {
+    private long followerId;
+    private long followeeId;
+}
