+++ src/main/java/faang/school/notificationservice/event/mentorship/MentorshipAcceptedEvent.java
@@ -1,0 +1,19 @@
+package faang.school.notificationservice.event.mentorship;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class MentorshipAcceptedEvent {
+
+    private long id;
+    private long requesterId;
+    private String receiverUsername;
+}
