+++ src/main/java/faang/school/notificationservice/listener/impl/AchievementEventListener.java
@@ -2,16 +2,15 @@ package faang.school.notificationservice.listener.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
-import faang.school.notificationservice.dto.event.AchievementEvent;
+import faang.school.notificationservice.event.achievement.AchievementEvent;
 import faang.school.notificationservice.listener.AbstractEventListener;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.Topic;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -41,7 +40,7 @@ public class AchievementEventListener extends AbstractEventListener<AchievementE
     }
 
     @Override
-    public Topic getTopic() {
+    public ChannelTopic getTopic() {
         return new ChannelTopic(topicName);
     }
 }
