+++ src/main/java/faang/school/notificationservice/listener/impl/CommentEventListener.java
@@ -5,11 +5,10 @@ import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.event.comment.CommentEventDto;
 import faang.school.notificationservice.listener.AbstractEventListener;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -18,7 +17,7 @@ import java.util.Locale;
 @Component
 public class CommentEventListener extends AbstractEventListener<CommentEventDto> {
 
-    @Value("${spring.data.redis.channels.comment}")
+    @Value("${spring.data.redis.channel.comment}")
     private String commentTopic;
 
     public CommentEventListener(ObjectMapper objectMapper,
@@ -26,11 +25,6 @@ public class CommentEventListener extends AbstractEventListener<CommentEventDto>
                                 List<NotificationService> notificationService,
                                 List<MessageBuilder<CommentEventDto>> messageBuilder) {
         super(objectMapper, userServiceClient, notificationService, messageBuilder);
-    }
-
-    @Override
-    public MessageListenerAdapter getAdapter() {
-        return new MessageListenerAdapter(this);
     }
 
     @Override
