+++ src/main/java/faang/school/notificationservice/listener/impl/EventStartEventListener.java
@@ -2,14 +2,13 @@ package faang.school.notificationservice.listener.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
-import faang.school.notificationservice.dto.EventStartEventDto;
+import faang.school.notificationservice.event.event_start.EventStartEventDto;
 import faang.school.notificationservice.listener.AbstractEventListener;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.Topic;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -39,7 +38,7 @@ public class EventStartEventListener extends AbstractEventListener<EventStartEve
     }
 
     @Override
-    public Topic getTopic() {
+    public ChannelTopic getTopic() {
         return new ChannelTopic(eventStartTopic);
     }
 }
