+++ src/main/java/faang/school/notificationservice/listener/impl/MentorshipAcceptedEventListener.java
@@ -2,15 +2,13 @@ package faang.school.notificationservice.listener.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
-import faang.school.notificationservice.event.MentorshipAcceptedEvent;
+import faang.school.notificationservice.event.mentorship.MentorshipAcceptedEvent;
 import faang.school.notificationservice.listener.AbstractEventListener;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.Topic;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -39,12 +37,7 @@ public class MentorshipAcceptedEventListener extends AbstractEventListener<Mento
     }
 
     @Override
-    public MessageListenerAdapter getAdapter() {
-        return new MessageListenerAdapter(this);
-    }
-
-    @Override
-    public Topic getTopic() {
+    public ChannelTopic getTopic() {
         return new ChannelTopic(mentorshipAcceptedChannel);
     }
 }
