+++ src/main/java/faang/school/notificationservice/listener/impl/UserFollowerEventListener.java
@@ -3,14 +3,12 @@ package faang.school.notificationservice.listener.impl;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.listener.AbstractEventListener;
-import faang.school.notificationservice.event.UserFollowerEvent;
+import faang.school.notificationservice.event.follower.UserFollowerEvent;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.Topic;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -40,12 +38,7 @@ public class UserFollowerEventListener extends AbstractEventListener<UserFollowe
     }
 
     @Override
-    public MessageListenerAdapter getAdapter() {
-        return new MessageListenerAdapter(this);
-    }
-
-    @Override
-    public Topic getTopic() {
+    public ChannelTopic getTopic() {
         return new ChannelTopic(channelName);
     }
 }
