+++ src/main/java/faang/school/notificationservice/messaging/impl/EventStartMessageBuilder.java
@@ -1,0 +1,27 @@
+package faang.school.notificationservice.messaging.impl;
+
+import faang.school.notificationservice.event.event_start.EventStartEventDto;
+import faang.school.notificationservice.messaging.MessageBuilder;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Service;
+
+import java.util.Locale;
+
+@Service
+@RequiredArgsConstructor
+public class EventStartMessageBuilder implements MessageBuilder<EventStartEventDto> {
+
+    private final MessageSource messageSource;
+
+    @Override
+    public Class<EventStartEventDto> supportEventType() {
+        return EventStartEventDto.class;
+    }
+
+    @Override
+    public String buildMessage(EventStartEventDto event, Locale locale) {
+        String message = "event." + event.getTimeBeforeStart();
+        return messageSource.getMessage(message, new Object[]{event.getEventId()}, locale);
+    }
+}
