+++ src/main/java/faang/school/notificationservice/messaging/impl/MentorshipAcceptedEventMessageBuilder.java
@@ -1,0 +1,28 @@
+package faang.school.notificationservice.messaging.impl;
+
+import faang.school.notificationservice.event.mentorship.MentorshipAcceptedEvent;
+import faang.school.notificationservice.messaging.MessageBuilder;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Component;
+
+import java.util.Locale;
+
+@Component
+@RequiredArgsConstructor
+public class MentorshipAcceptedEventMessageBuilder implements MessageBuilder<MentorshipAcceptedEvent> {
+
+    private static final String MESSAGE_KEY = "mentorship.accepted";
+
+    private final MessageSource messageSource;
+
+    @Override
+    public Class<MentorshipAcceptedEvent> supportEventType() {
+        return MentorshipAcceptedEvent.class;
+    }
+
+    @Override
+    public String buildMessage(MentorshipAcceptedEvent event, Locale locale) {
+        return messageSource.getMessage(MESSAGE_KEY, new Object[] {event.getReceiverUsername()}, locale);
+    }
+}
