--- src/main/java/faang/school/notificationservice/service/NotificationService.java
@@ -1,10 +1,0 @@
-package faang.school.notificationservice.service;
-
-import faang.school.notificationservice.dto.UserDto;
-
-public interface NotificationService {
-
-    void send(UserDto user, String message);
-
-    UserDto.PreferredContact getPreferredContact();
-}
