+++ src/main/java/faang/school/notificationservice/service/notification/NotificationService.java
@@ -1,0 +1,10 @@
+package faang.school.notificationservice.service.notification;
+
+import faang.school.notificationservice.dto.UserDto;
+
+public interface NotificationService {
+
+    void send(UserDto user, String message);
+
+    UserDto.PreferredContact getPreferredContact();
+}
