+++ src/main/java/faang/school/notificationservice/service/notification/impl/email/EmailService.java
@@ -1,0 +1,28 @@
+package faang.school.notificationservice.service.notification.impl.email;
+
+import faang.school.notificationservice.dto.UserDto;
+import faang.school.notificationservice.service.notification.NotificationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class EmailService implements NotificationService {
+
+    private final JavaMailSender emailSender;
+
+    @Override
+    public void send(UserDto user, String message) {
+        SimpleMailMessage mailMessage = new SimpleMailMessage();
+        mailMessage.setTo(user.getEmail());
+        mailMessage.setText(message);
+        emailSender.send(mailMessage);
+    }
+
+    @Override
+    public UserDto.PreferredContact getPreferredContact() {
+        return UserDto.PreferredContact.EMAIL;
+    }
+}
