+++ src/test/java/faang/school/notificationservice/listener/EventListenerForTest.java
@@ -3,10 +3,9 @@ package faang.school.notificationservice.listener;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.messaging.MessageBuilder;
-import faang.school.notificationservice.service.NotificationService;
+import faang.school.notificationservice.service.notification.NotificationService;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.Topic;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 
 import java.util.List;
@@ -21,22 +20,12 @@ public class EventListenerForTest extends AbstractEventListener<EventForTest> {
     }
 
     @Override
-    public void onMessage(Message message, byte[] pattern) {
-
-    }
-
-    @Override
     public MessageListenerAdapter getAdapter() {
         return null;
     }
 
     @Override
     public ChannelTopic getTopic() {
-        return null;
-    }
-
-    @Override
-    public Topic getTopic() {
         return new ChannelTopic("test_topic");
     }
